add crossfade between songs
Liquidsoap likes to spit out a warning when this option is used, but it
seems to work fine for at least several hours.

Warning: #get called when not #is_ready! This is normal if an operator
using this source has been unused while the source has gone unavailable.
If unsure about this warning, you are very welcome to report it and ask
for clarifications.