Asynchronously log metadata
I recently looked into a silence alarm that went out to the engineering
staff and realized that it was legitimate and was being caused by an
issue with johnny-six. Namely, I missed the fact that on_track runs in
the main streaming thread...so any latency in logging the track can
actually delay playback on johnny-six. We obviously don't want that, so
I've used add_timeout so that the log_metadata calls run in a separate
thread now.