Add "transition_with_skip" that skips track
This is used for the transition to the radio source. This way, when a
prerecorded show ends, the switch statement that determines what to play
will be called again to ensure traffic is played as quickly as possible.
This may also help with ensuring that we aren't playing music outside of
its scheduled hour.